# docs/swagger.yaml
definitions:
...
  dto.ReturnPolicyCreateReq:
    properties:
      accepts_exchanges:
        type: boolean
      accepts_returns:
        type: boolean
      return_deadline:
        minimum: 0
        type: integer
      shop_id:
        type: integer
    required:
    - shop_id
    type: object
  dto.ReturnPolicyListResp:
    properties:
      list:
        items:
          $ref: '#/definitions/dto.ReturnPolicyResp'
        type: array
      total:
        type: integer
    type: object
  dto.ReturnPolicyResp:
    properties:
      accepts_exchanges:
        type: boolean
      accepts_returns:
        type: boolean
      created_at:
        type: string
      etsy_policy_id:
        type: integer
      etsy_synced_at:
        type: string
      id:
        type: integer
      return_deadline:
        type: integer
      shop_id:
        type: integer
      updated_at:
        type: string
    type: object
  dto.ReturnPolicyUpdateReq:
    properties:
      accepts_exchanges:
        type: boolean
      accepts_returns:
        type: boolean
      return_deadline:
        minimum: 0
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/return-policies/{id}:
    delete:
      description: 从 Etsy 删除退货政策
      parameters:
      - description: 退货政策ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "删除成功"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: ID格式错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 删除失败
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除退货政策
      tags:
      - ReturnPolicy (退货政策)
    get:
      description: 获取退货政策的详细配置信息
      parameters:
      - description: 退货政策ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 退货政策详情
          schema:
            $ref: '#/definitions/dto.ReturnPolicyResp'
        "400":
          description: ID格式错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 查询失败
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取退货政策详情
      tags:
      - ReturnPolicy (退货政策)
    put:
      consumes:
      - application/json
      description: 更新 Etsy 退货政策配置
      parameters:
      - description: 退货政策ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ReturnPolicyUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "更新成功"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 更新失败
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新退货政策
      tags:
      - ReturnPolicy (退货政策)
...
  /api/v1/shops/{shopId}/return-policies:
    get:
      description: 获取指定店铺的所有退货政策
      parameters:
      - description: 店铺ID
        in: path
        name: shopId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 退货政策列表
          schema:
            $ref: '#/definitions/dto.ReturnPolicyListResp'
        "400":
          description: ID格式错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 查询失败
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取退货政策列表
      tags:
      - ReturnPolicy (退货政策)
    post:
      consumes:
      - application/json
      description: 在 Etsy 创建新的退货政策
      parameters:
      - description: 店铺ID
        in: path
        name: shopId
        required: true
        type: integer
      - description: 政策参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ReturnPolicyCreateReq'
      produces:
      - application/json
      responses:
        "201":
          description: 创建结果
          schema:
            $ref: '#/definitions/dto.ReturnPolicyResp'
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 创建失败
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建退货政策
      tags:
      - ReturnPolicy (退货政策)
  /api/v1/shops/{shopId}/return-policies/sync:
    post:
      description: 从 Etsy 同步店铺的退货政策数据
      parameters:
      - description: 店铺ID
        in: path
        name: shopId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "退货政策同步成功"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: ID格式错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 同步失败
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 同步退货政策
      tags:
      - ReturnPolicy (退货政策)
...
swagger: "2.0"